Refactor codebase to modernize
This is a relatively large cleanup of the codebase to try and bring it
closer to modern Python development practices. In addition to pruning
old files and switching to Pipenv, this update changes the way the
database is interacted with. The main goal with this portion of the
rewrite was to allow for connection reuse across requests.