Clean up authorization
This moves the auth config into the rest of the app config. It also
cleans up the logic a bit to return a 500 status code if the server auth
token has not been set and default to a 401 status code unless the token
strings match. The string comparison has been switched to a constant
time algorithm for security.