Optimize the Lambda package size
The Lambda package that zappa builds is right on the borderline of
being too big to deploy. Because Travis by default includes a bunch of
garbage (specifically numpy) in the base python install, this was
pushing the package over the limit and it was failing to deploy. The
current incantation is the best way I could find to create a clean
Lambda deploy package.