Use connection pooling
The app was not originally written to use database connection pooling.
This was not a problem when it was being run as a lambda since the
process (and connection) was short lived. When it got moved to a more
traditional EC2 deployment this became a problem.

This commit just swaps out the single long lived connection for a pool.